Factor logging for tests out to a separate file
